+++ dakaraphi.devtools.tracing/src/main/java/dakaraphi/devtools/tracing/TracingAgent.java
@@ -5,11 +5,15 @@ import java.io.IOException;
 import java.lang.instrument.Instrumentation;
 import java.lang.reflect.InvocationTargetException;
 
+import com.fasterxml.jackson.core.JsonParseException;
+import com.fasterxml.jackson.databind.JsonMappingException;
+
 import dakaraphi.devtools.tracing.config.ConfigurationSerializer;
 import dakaraphi.devtools.tracing.config.TracingConfig;
 import dakaraphi.devtools.tracing.filewatcher.FileWatcher;
 import dakaraphi.devtools.tracing.filewatcher.IFileListener;
 import dakaraphi.devtools.tracing.logger.TraceLogger;
+import dakaraphi.devtools.tracing.metrics.ExecutionCounts;
 
 /**
  * TracingAgent can only be launched within a jar file.
@@ -29,25 +33,50 @@ import dakaraphi.devtools.tracing.logger.TraceLogger;
 public class TracingAgent {
 	public static TracingConfig tracingConfig = null;
     public static void premain(String agentArgs, Instrumentation instrumentation) throws IOException, ClassNotFoundException, NoSuchMethodException, SecurityException, IllegalAccessException, IllegalArgumentException, InvocationTargetException, InstantiationException {
-    	TraceLogger.log("Starting v0.5.1");
-       	TraceLogger.log("classloader: " + TracingAgent.class.getClassLoader());
+    	TraceLogger.log("Starting v1.0.8");
+		String tracerDefinitionFile = System.getProperty(ConfigurationSerializer.FILE_PROPERTY_KEY);
+		if (tracerDefinitionFile == null) {
+			TraceLogger.log("Missing system property " + ConfigurationSerializer.FILE_PROPERTY_KEY);
+			return;
+		}
 		final File configFile = new File(System.getProperty(ConfigurationSerializer.FILE_PROPERTY_KEY));
-		ClassMethodSelector selector = loadConfig(configFile);
+		TraceLogger.log("using tracer definitions: " + configFile);
+
+		ClassMethodSelector selector = null;
+		try {
+			selector = loadConfig(configFile);
+		} catch (Throwable e) {
+			TraceLogger.log("unable to load configuration definition - " + e.getMessage());
+		}
     	
     	final TracingTransformer transformer = new TracingTransformer(instrumentation, selector, new MethodRewriter());
     	
     	FileWatcher.createFileWatcher().addListener(new IFileListener() {
-			public void onFileChange() {
+			public void onFileChange() throws Throwable {
 				ClassMethodSelector selector = loadConfig(configFile);
 				transformer.setClassMethodSelector(selector);
 				transformer.retransform();
+				StacktraceHasher.clear();
+				ExecutionCounts.clear();
 			}
 		}, configFile).start();
 	}
 	
-	private static ClassMethodSelector loadConfig(File configFile) {
+	private static ClassMethodSelector loadConfig(File configFile) throws JsonParseException, JsonMappingException, IOException {
 		ConfigurationSerializer serializer = new ConfigurationSerializer(configFile);
 		tracingConfig = serializer.readConfig();
 		return ClassMethodSelector.makeSelector(tracingConfig);
 	}
 }
+
+/*
+ TODO 
+ - consider exposing a live interface using - https://github.com/perwendel/spark
+ - add support for triggers
+ - add support for constructors
+ - allow tracers to override the global logConfig
+ - add support for timers.  Just measure and log diff between each occurrence of timer by same name. option, on same thread.
+ - tracing a low level function like System.out.println.  The class will have already been loaded when the agent starts and therefore will not invoke the transformer
+	- it will get invoked if the tracer file is touched.  Need to investigate if there is a way to detect at agent startup.
+*/
+