+++ dakaraphi.devtools.tracing/src/main/java/dakaraphi/devtools/tracing/config/Tracer.java
@@ -14,10 +14,21 @@ public class Tracer {
 	public String name;
 	public Pattern classRegex;
 	public Pattern methodRegex;
-	public Integer line;
-	public List<String> variables;
+	public List<Type> types = new ArrayList<>();
+	public Integer line = Integer.valueOf(0);
+	public List<Variable> variables = new ArrayList<>();
 	public LogWhen logWhen;
 	public LogStackFrames logStackFrames;
+
+	public static class Variable {
+		public String expression;
+		public String name;
+	}
+
+	public static class Type {
+		public Integer index;
+		public Pattern typeRegex;
+	}
 
 	public static class LogWhen {
 		public List<VariableCondition> variableValues = new ArrayList<>();
@@ -26,7 +37,7 @@ public class Tracer {
 	}
 
 	public static class VariableCondition {
-		public int index;
+		public String name;
 		public Pattern valueRegex;
 	}
 
@@ -34,5 +45,6 @@ public class Tracer {
 		public int limit;
 		public Pattern includeRegex;
 		public Pattern excludeRegex;
+		public boolean referenceDuplicatesByHash;
 	}
 }