+++ dakaraphi.devtools.tracing/src/test/java/dakaraphi/devtools/PlainObject.java
@@ -1,0 +1,7 @@
+package dakaraphi.devtools;
+
+public class PlainObject {
+    PlainObject() {
+        System.out.println("invoked constructor PlainObject");
+    }
+}